+++ main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/zimeg/instant-band-night/cmd"
+)
+
+// main begins the execution of this program
+func main() {
+	cmd.Execute()
+}